MB-59751: Make saml assertion dupe check more strict
Check if assertion (which part of the response message) is being
replayed instead of checking the whole response.
If we check the whole response it still leaves a possibility to
replay the assertion by inserting it inside of slightly modified
response (If assertion is signed and reponse is not signed it can
be easily done). This makes the response message digest different
and the dupe check will not notice that the assertion is being
replayed.

Change-Id: I1bb17e1fd4183d40eb466c1eda81e65d2621fcde
Reviewed-on: https://review.couchbase.org/c/ns_server/+/201923
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>