MB-59751: Use assertion ID for dupe check...
... instead of assertion digest

Quote from the saml threat modeling document:
"Assertions should contain a unique ID that is only accepted once
 by the application"

Using digest is not always reliable. In some cases it is possible
to generate different digest for absolutely same assertions (the
difference sits somewhere in the response message that contains
the assertion, but what exactly affects the assertion diggest in
this case remains a secret to me).

It also seems like there are other implementations that also use ID
for dupe checks.

Note also that changing assertion ID must always affect the assertion
digest, which should lead to signature validation failure. And that
means that by making this change we are only making the dupe check
more strict.

Change-Id: I352127720afbd97a86dca251186b5aac295a777e
Reviewed-on: https://review.couchbase.org/c/ns_server/+/202763
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Navdeep S Boparai <navdeep.boparai@couchbase.com>
Tested-by: Timofey Barmin <timofey.barmin@couchbase.com>